Extract public pad create flow
