+++ js/etherpad_nextcloud-files-main.mjs
@@ -1,2 +1,2 @@
-import{a as T,A as p,o as de,b as ge,c as ve,d as X,n as Ce,e as Ee,M as Y,f as xe}from"./oc-compat-Cn8-bqPr.chunk.mjs";import{p as U,a as z,v as D,f as j,b as Ae,r as Z,c as ue,g as K,n as R,i as B,d as Ne,e as Le}from"./urls-D-89WmZ3.chunk.mjs";const P=new Map,$=async n=>{const i=String(n);if(P.has(i))return P.get(i);const e=T("/apps/"+p+"/api/v1/pads/resolve")+"?fileId="+encodeURIComponent(i),r=_(e,{method:"GET",headers:{Accept:"application/json"}},"Pad resolve failed.").catch(a=>{throw P.delete(i),a});return P.set(i,r),r},Se=async n=>{const i="path:"+String(n);if(P.has(i))return P.get(i);const e=T("/apps/"+p+"/api/v1/pads/resolve")+"?file="+encodeURIComponent(n),r=_(e,{method:"GET",headers:{Accept:"application/json"}},"Pad resolve by path failed.").catch(a=>{throw P.delete(i),a});return P.set(i,r),r},V=async n=>{const i=T("/apps/"+p+"/api/v1/pads/sync-status/"+encodeURIComponent(String(n)));return _(i,{method:"GET",headers:{Accept:"application/json"}},"Sync status check failed.")},Oe=async(n,i)=>{let e=T("/apps/"+p+"/api/v1/pads/sync/"+encodeURIComponent(String(n)));return e+="?force=1",_(e,{method:"POST",headers:{Accept:"application/json",requesttoken:de()}},"Pad sync failed.")},Ie=async(n,i)=>{const e=T("/apps/"+p+"/api/v1/pads/from-url"),r=new URLSearchParams;return r.set("file",n),r.set("padUrl",i),_(e,{method:"POST",headers:pe(),body:r.toString()},"Could not import public pad URL.")},Me=async n=>{const i=T("/apps/"+p+"/api/v1/pads"),e=new URLSearchParams;return e.set("file",n),e.set("accessMode","public"),_(i,{method:"POST",headers:pe(),body:e.toString()},"Could not create public pad.")},pe=()=>({Accept:"application/json","Content-Type":"application/x-www-form-urlencoded;charset=UTF-8",requesttoken:de()}),_=async(n,i,e)=>{const r=await fetch(n,{...i,credentials:"same-origin"}),a=await r.json().catch(()=>({}));if(!r.ok)throw new Error(a&&a.message||e);return a},q=()=>!!(window.OCA&&window.OCA.Viewer&&typeof window.OCA.Viewer.open=="function"),k=()=>(window.location.pathname||"").includes("/apps/files"),ke=120,Pe=900,ee=async(n,i={})=>{if(q()&&n.path)try{const e=Number(n.fileId);k()&&Number.isFinite(e)&&e>0&&(_e(e,n.path,i.resolveOpenDir),await Te()),await Fe(n.path),window.OCA.Viewer.open({path:n.path,onClose:He});return}catch{}typeof i.fallbackOpen=="function"&&await i.fallbackOpen(n)},Te=()=>new Promise(n=>{window.setTimeout(n,ke)}),Fe=async n=>{const i=ge(n);i!==""&&(ve("editor:file:created",i),await qe())},qe=()=>new Promise(n=>{window.setTimeout(n,Pe)}),_e=(n,i,e)=>{const r=window.OCP&&window.OCP.Files&&window.OCP.Files.Router;!r||typeof r.goToRoute!="function"||r.goToRoute(null,{...r.params||{},view:r.params&&r.params.view?r.params.view:"files",fileid:String(n)},{...r.query||{},dir:Re(i,e),editing:"false",openfile:"true"},!0)},Re=(n,i)=>{if(typeof i=="function")return i(n);const e=String(n||""),r=e.lastIndexOf("/");return r>0?e.substring(0,r):"/"},He=()=>{const n=window.OCP&&window.OCP.Files&&window.OCP.Files.Router;if(!n||typeof n.goToRoute!="function")return;const i={...n.query||{}};delete i.openfile,delete i.editing,n.goToRoute(null,n.params||{},i)},Ue=800,ze=180,Be=(n,i)=>{const e=window.OCP&&window.OCP.Files&&window.OCP.Files.Router;return!e||typeof e.goToRoute!="function"||!n||!Number.isFinite(n)?!1:(e.goToRoute(null,{view:"files",fileid:String(n)},{dir:Z(i),editing:"false",openfile:"true"}),window.setTimeout(()=>{if((window.location.pathname||"").includes("/apps/files/files/"+String(n)))return;const r=j(n,i);window.location.assign(r)},ze),!0)},De=()=>{let n=null,i=0;return async e=>{const r=String(e.fileId??"")+"|"+String(e.path??""),a=Date.now();if(n===r&&a-i<Ue)return;n=r,i=a;const s=U(),E=s!==null&&s!=="",f=()=>{if(E){window.location.assign(D(s,e.path||""));return}if(k()&&e.fileId!==null&&e.fileId!==void 0&&Number.isFinite(Number(e.fileId))){const N=e.path||"/";window.location.assign(j(Number(e.fileId),N));return}const g=k()?z():null;if((e.fileId===null||e.fileId===void 0)&&g){const N=e.path||"/";window.location.assign(j(g,N));return}if(e.fileId!==null&&e.fileId!==void 0&&Number.isFinite(Number(e.fileId))){const N=e.path||"/";window.location.assign(j(Number(e.fileId),N));return}e.path&&window.location.assign(Ae(e.path))};if(!q()){f();return}let y=e.path||"",w=e.fileId??null;if(!y&&e.fileId!==null&&e.fileId!==void 0)try{const g=await $(e.fileId);y=g&&typeof g.path=="string"?g.path:"",w=g&&Number.isFinite(Number(g.file_id))?Number(g.file_id):w}catch{y=""}if(!y){f();return}if((!w||!Number.isFinite(Number(w)))&&y&&!E)try{const g=await Se(y);w=g&&Number.isFinite(Number(g.file_id))?Number(g.file_id):w}catch{}if((!w||!Number.isFinite(Number(w)))&&k()){const g=z();g&&(w=g)}if(k()){if(w&&Number.isFinite(Number(w))&&Be(Number(w),y))return;f();return}try{window.OCA.Viewer.open({path:y})}catch{f()}}},G="data-epnc-sidebar-sync-panel",Q="data-epnc-sidebar-sync-mount",fe=8e3,je=6e4,me=15e3,$e=8e3,We=()=>{const n=document.documentElement,i=document.body,e=(n?.className||"")+" "+(i?.className||"");if(/theme[-_]?dark|dark-theme|theme--dark/i.test(e))return!0;const r=window.matchMedia&&window.matchMedia("(prefers-color-scheme: dark)");return!!(r&&r.matches)},te=n=>{if(!(n instanceof HTMLElement))return!1;const i=window.getComputedStyle(n);return i.display==="none"||i.visibility==="hidden"?!1:i.position==="fixed"?!0:n.offsetParent!==null},Ke=n=>{const i=Number(n);return Number.isFinite(i)&&i>0?i:null},Ve=()=>{const n=z();if(n!==null)return n;const i=new URLSearchParams(window.location.search||""),e=Number(i.get("fileid")||"");return Number.isFinite(e)&&e>0?e:null},ne=()=>{const n=["[data-cy-files-sidebar]","#app-sidebar-vue .app-sidebar","#app-sidebar-vue",".app-sidebar"];for(const i of n){const e=document.querySelector(i);if(e instanceof HTMLElement)return e}return null},Ge=n=>{if(!(n instanceof HTMLElement))return null;let i=n.querySelector("["+Q+'="1"]');i instanceof HTMLElement||(i=document.createElement("div"),i.className="epnc-sidebar-sync-mount",i.setAttribute(Q,"1"));const e=n.querySelector("[data-cy-files-sidebar-tabs], .app-sidebar-tabs");if(e&&e.parentNode instanceof HTMLElement){const r=e.parentNode;return(i.parentNode!==r||i.nextSibling!==e)&&r.insertBefore(i,e),i}return i.parentNode!==n&&n.insertBefore(i,n.firstChild),i},ie=n=>{const i=Ve();if(i!==null)return i;if(!(n instanceof HTMLElement))return null;const e=["[data-fileid]","[data-file-id]","[data-node-id]","[data-id]"];for(const a of e){const s=n.querySelector(a);if(!(s instanceof HTMLElement))continue;const E=[s.getAttribute("data-fileid"),s.getAttribute("data-file-id"),s.getAttribute("data-node-id"),s.getAttribute("data-id"),s.dataset&&(s.dataset.fileid||s.dataset.fileId)];for(const f of E){const y=Ke(f);if(y!==null)return y}}const r=n.querySelectorAll("a[href]");for(const a of r){if(!(a instanceof HTMLAnchorElement))continue;const s=ue(a.getAttribute("href"));if(s!==null)return s}return null},Je=n=>n==="synced"?"etherpad-icon-color":We()?"etherpad-icon-white":"etherpad-icon-black",Ze=n=>n==="synced"?t(p,"Synchronized"):n==="syncing"?t(p,"Saving..."):n==="error"?t(p,"Sync failed"):n==="unavailable"?t(p,"Sync status unavailable"):n==="loading"?t(p,"Checking sync status..."):t(p,"Sync pending"),J=n=>n==="synced"?"synced":n==="out_of_sync"?"pending":n==="unavailable"?"unavailable":"error",Qe=(n,i)=>n==="synced"?Math.min(je,i+$e):n==="pending"?fe:me,Xe=()=>{let n=null,i=null,e=null,r=0;const a=()=>{e!==null&&(window.clearTimeout(e),e=null)},s=()=>{const d=document.querySelector("["+G+'="1"]');d&&d.parentNode&&d.parentNode.removeChild(d);const u=document.querySelector("["+Q+'="1"]');u&&u.parentNode&&u.parentNode.removeChild(u),a()},E=(d,u)=>{if(!(d instanceof HTMLElement))return;const c=String(u||"").trim();d.setAttribute("data-public-open-url",c);const C=d.querySelector("[data-epnc-sidebar-open-public-button]");if(!(C instanceof HTMLButtonElement))return;const A=c!=="";C.style.display=A?"inline-flex":"none",C.disabled=!A},f=(d,u,c,C,A=!0)=>{if(!(d instanceof HTMLElement))return;const x=d.querySelector("[data-epnc-sidebar-sync-icon]"),S=d.querySelector("[data-epnc-sidebar-sync-label]"),L=d.querySelector("[data-epnc-sidebar-sync-button]");if(d.setAttribute("data-file-id",String(c)),d.setAttribute("data-sync-state",u),d.setAttribute("data-sync-in-flight",C?"1":"0"),S instanceof HTMLElement&&(S.textContent=Ze(u)),x instanceof HTMLElement){const I=Je(u),o=X(p,I);o!==""&&(x.style.backgroundImage=`url(${o})`)}L instanceof HTMLButtonElement&&(L.disabled=C||!A||c<=0,L.setAttribute("aria-busy",C?"true":"false"))},y=(d,u)=>{if(!(u instanceof HTMLElement))return null;let c=u.querySelector("["+G+'="1"]');if(!(c instanceof HTMLElement)){c=document.createElement("section"),c.className="epnc-sidebar-sync-panel",c.setAttribute(G,"1"),c.setAttribute("data-file-id",String(d));const C=document.createElement("div");C.className="epnc-sidebar-sync-panel__header";const A=document.createElement("span");A.className="epnc-sidebar-sync-panel__icon",A.setAttribute("data-epnc-sidebar-sync-icon","1"),A.setAttribute("aria-hidden","true");const x=document.createElement("span");x.className="epnc-sidebar-sync-panel__status",x.setAttribute("data-epnc-sidebar-sync-label","1"),C.appendChild(A),C.appendChild(x);const S=document.createElement("button");S.type="button",S.className="button-vue button-vue--size-normal button-vue--vue-secondary epnc-sidebar-sync-panel__button",S.setAttribute("data-epnc-sidebar-sync-button","1"),S.textContent=t(p,"Pad in Datei speichern"),S.addEventListener("click",async I=>{I.preventDefault(),I.stopPropagation();const o=Number(c.getAttribute("data-file-id")||"");if(!(!Number.isFinite(o)||o<=0)){f(c,"syncing",o,!0,!0);try{await Oe(o,!0);const l=await V(o),m=J(l&&l.status);f(c,m,o,!1,!0)}catch{f(c,"error",o,!1,!0)}}});const L=document.createElement("button");L.type="button",L.className="button-vue button-vue--size-normal button-vue--vue-tertiary epnc-sidebar-sync-panel__button epnc-sidebar-sync-panel__button--public-open",L.setAttribute("data-epnc-sidebar-open-public-button","1"),L.textContent=t(p,"Pad in neuem Tab öffnen"),L.style.display="none",L.addEventListener("click",I=>{I.preventDefault(),I.stopPropagation();const o=String(c.getAttribute("data-public-open-url")||"").trim();o!==""&&window.open(o,"_blank","noopener,noreferrer")}),c.appendChild(C),c.appendChild(S),c.appendChild(L)}return c.parentNode!==u&&u.appendChild(c),c},w=(d,u)=>{if(a(),!(d instanceof HTMLElement)||!Number.isFinite(u)||u<=0)return;let c=fe;const C=x=>{e=window.setTimeout(()=>{A()},Math.max(1e3,x))},A=async()=>{if(!(d instanceof HTMLElement)||!document.body.contains(d)){a();return}const x=ne();if(!(x instanceof HTMLElement)||!te(x)||!te(d)){a();return}if(ie(x)!==u){a();return}if(document.visibilityState!=="visible"||d.getAttribute("data-sync-in-flight")==="1"){C(c);return}try{const S=await V(u),L=J(S&&S.status);f(d,L,u,!1,!0),c=Qe(L,c)}catch{f(d,"error",u,!1,!0),c=me}C(c)};C(c)},g=()=>{i!==null&&(window.clearTimeout(i),i=null)},N=async()=>{const d=++r;if(!k()||U()){s();return}const u=ne();if(!(u instanceof HTMLElement)){s();return}const c=ie(u);if(!c){s();return}const C=Ge(u);if(!(C instanceof HTMLElement)){s();return}let A;try{A=await $(c)}catch{s();return}if(d!==r)return;if(!A||A.is_pad!==!0){s();return}const x=y(c,C);if(!(x instanceof HTMLElement)){s();return}const S=A&&typeof A.public_open_url=="string"?A.public_open_url.trim():"";E(x,S),f(x,"loading",c,!1,!0);try{const L=await V(c);if(d!==r)return;const I=J(L&&L.status);f(x,I,c,!1,!0),w(x,c)}catch{if(d!==r)return;f(x,"error",c,!1,!0),w(x,c)}},v=(d=0)=>{g(),i=window.setTimeout(()=>{i=null,N()},Math.max(0,d))};return{installObserver:()=>{if(n!==null||!k())return;const d=new MutationObserver(u=>{for(const c of u)for(const C of c.addedNodes)if(C instanceof Element&&(C.matches("[data-cy-files-sidebar], #app-sidebar-vue, .app-sidebar")||C.querySelector("[data-cy-files-sidebar], #app-sidebar-vue, .app-sidebar"))){v(60);return}});d.observe(document.body,{childList:!0,subtree:!0}),n=d},scheduleRefresh:v}},re=p+"_public_pad",oe="icon-filetype-etherpad-nextcloud-pad",ae='<svg xmlns="http://www.w3.org/2000/svg" xml:space="preserve" fill-rule="evenodd" stroke-linejoin="round" stroke-miterlimit="2" clip-rule="evenodd" viewBox="0 0 355 355"><path fill="#5ac395" d="M317 89v177c0 28-23 51-51 51H89c-28 0-51-23-51-51V89c0-28 23-51 51-51h177c28 0 51 23 51 51"/><path fill="#4aa57f" fill-rule="nonzero" d="M240 144q-4-3-8 0-6 6 0 9a36 36 0 0 1 0 52 6 6 0 0 0 3 12q3 0 5-3c19-18 19-50 0-70"/><path fill="#479a76" fill-rule="nonzero" d="M267 122q-4-5-9 0-4 4 0 10c25 26 25 68 0 93q-4 5 0 9 5 4 9 0c30-31 30-81 0-112"/><path fill="#fff" d="M192 130v2q-1 8-10 10H76q-9-2-11-10v-2q1-10 11-10h106q9 0 10 10m24 51v2q-1 11-12 11H78q-10 0-12-11v-2q2-10 12-11h126q12 1 12 11"/><path fill="#fff" d="M216 181v2q-1 11-12 11H78q-10 0-12-11v-2q2-10 12-11h126q12 1 12 11m-57 52v1q-1 11-10 11H76q-9 0-11-11v-1q1-10 11-11h73q9 1 10 11"/><path fill="#fff" d="M159 233v1q-1 11-10 11H76q-9 0-11-11v-1q1-10 11-11h73q9 1 10 11"/></svg>',se=98,Ye=120,et=500,le=()=>{const n=Ce();return n!==null&&n>=33},H=n=>(n&&n.message?String(n.message):"").toLowerCase().includes("duplicate"),tt=n=>{if(!n||typeof n!="object")return null;if(n.hasCreatePermission===!1)return!1;if(n.hasCreatePermission===!0)return!0;if(typeof n.canCreate=="boolean")return n.canCreate;const i=Ee(),e=[n.permissions,n.permission,n.attributes&&n.attributes.permissions,typeof n.get=="function"?n.get("permissions"):null];for(const r of e){const a=Number(r);if(Number.isFinite(a))return(a&i)===i}return null},nt=(...n)=>{for(const i of n){const e=tt(i);if(e!==null)return e}return!0},it=()=>{const n=[window.OCP&&window.OCP.Files,window.OCA&&window.OCA.Files];for(const e of n)if(!(!e||typeof e!="object")){if(typeof e.addNewFileMenuEntry=="function")return{register:e.addNewFileMenuEntry.bind(e),unregister:typeof e.removeNewFileMenuEntry=="function"?e.removeNewFileMenuEntry.bind(e):null};if(typeof e.getNewFileMenu=="function"){let r=null;try{r=e.getNewFileMenu()}catch{r=null}if(r&&typeof r.registerEntry=="function")return{register:r.registerEntry.bind(r),unregister:typeof r.unregisterEntry=="function"?r.unregisterEntry.bind(r):null}}}const i=window._nc_newfilemenu;return i&&typeof i=="object"&&typeof i.registerEntry=="function"?{register:i.registerEntry.bind(i),unregister:typeof i.unregisterEntry=="function"?i.unregisterEntry.bind(i):null}:null},rt=({isFilesAppRoute:n,onCreatePublicPad:i})=>{let e=!1,r=!1,a=!1,s=0;const E=()=>{i()},f=(...v)=>nt(...v),y=()=>({id:re,displayName:t(p,"Public pad"),...le()?{iconSvgInline:ae}:{iconClass:oe},order:se,enabled:f,handler:E}),w=()=>({id:re,displayName:t(p,"Public pad"),templateName:t(p,"Public pad"),iconClass:oe,...le()?{iconSvgInline:ae}:{},fileType:"file",order:se,actionHandler:E}),g=async()=>{const v=it();if(!v)return!1;const d=y();try{return v.register(d),e=!0,!0}catch(u){return H(u)?(e=!0,!0):!1}},N=()=>{if(r)return!0;const v=w(),d=()=>{const u=[window.OCA&&window.OCA.Files&&window.OCA.Files.NewFileMenu,window.OCA&&window.OCA.Files&&window.OCA.Files.newFileMenu];for(const c of u)if(!(!c||typeof c!="object")){if(typeof c.addMenuEntry=="function")try{return c.addMenuEntry(v),!0}catch(C){if(H(C))return!0}if(typeof c.registerEntry=="function")try{return c.registerEntry(y()),!0}catch(C){if(H(C))return!0}}return!1};try{return d()?(r=!0,!0):(!a&&window.OC&&window.OC.Plugins&&typeof window.OC.Plugins.register=="function"&&(window.OC.Plugins.register("OCA.Files.NewFileMenu",{attach(u){if(!(!u||typeof u.addMenuEntry!="function"))try{u.addMenuEntry(v),r=!0}catch(c){H(c)&&(r=!0)}}}),a=!0),r)}catch(u){return H(u)?(r=!0,!0):!1}};return()=>{if(!n()||e||r)return;const v=++s,d=async u=>{v!==s||e||r||await g()||N()||u<Ye&&window.setTimeout(()=>{d(u+1)},et)};d(0)}},ot=n=>{try{const i=new URL(n,window.location.origin),e=decodeURIComponent(i.pathname||"").match(/\/p\/([^/]+)$/);if(e&&e[1]){const r=e[1].replace(/[^a-zA-Z0-9._-]+/g,"-").replace(/^-+|-+$/g,"");if(r!=="")return r+".pad"}}catch{}return"Imported pad.pad"},ye=n=>{const i=document.createElement("div");i.style.position="fixed",i.style.inset="0",i.style.background="rgba(0, 0, 0, 0.45)",i.style.display="flex",i.style.alignItems="center",i.style.justifyContent="center",i.style.zIndex="20000";const e=document.createElement("div");e.style.position="relative",e.style.background="var(--color-main-background, #fff)",e.style.borderRadius="10px",e.style.boxShadow="0 10px 30px rgba(0, 0, 0, 0.25)",e.style.padding="18px",e.style.width="min(460px, calc(100vw - 24px))";const r=document.createElement("button");r.type="button",r.setAttribute("aria-label",t(p,"Close")),r.textContent="×",r.style.position="absolute",r.style.top="8px",r.style.right="10px",r.style.border="none",r.style.background="transparent",r.style.fontSize="22px",r.style.cursor="pointer",r.style.lineHeight="1";const a=document.createElement("h3");return a.textContent=n,a.style.margin="0 26px 10px 0",a.style.fontSize="18px",e.appendChild(r),e.appendChild(a),i.appendChild(e),document.body.appendChild(i),{overlay:i,dialog:e,closeButton:r}},at=()=>new Promise(n=>{const{overlay:i,dialog:e,closeButton:r}=ye(t(p,"Internal public pad")),a=document.createElement("label");a.textContent=t(p,"File name"),a.style.display="block",a.style.marginBottom="6px";const s=document.createElement("input");s.type="text",s.value=t(p,"Public pad")+".pad",s.style.width="100%",s.style.boxSizing="border-box",s.style.marginBottom="12px";const E=document.createElement("p");E.style.color="var(--color-error, #c62828)",E.style.margin="0 0 12px 0",E.style.minHeight="20px";const f=document.createElement("button");f.type="button",f.className="primary",f.textContent=t(p,"Create");const y=w=>{i.remove(),n(w)};r.addEventListener("click",()=>y(null)),i.addEventListener("click",w=>{w.target===i&&y(null)}),s.addEventListener("keydown",w=>{w.key==="Enter"&&(w.preventDefault(),f.click())}),f.addEventListener("click",()=>{const w=s.value.trim();if(w===""){E.textContent=t(p,"File name is required.");return}y(w)}),e.appendChild(a),e.appendChild(s),e.appendChild(E),e.appendChild(f),s.focus(),s.select()}),st=()=>new Promise(n=>{const{overlay:i,dialog:e,closeButton:r}=ye(t(p,"External public pad (URL)")),a=document.createElement("label");a.textContent=t(p,"External pad URL"),a.style.display="block",a.style.marginBottom="6px";const s=document.createElement("input");s.type="url",s.value="https://",s.placeholder="https://",s.style.width="100%",s.style.boxSizing="border-box",s.style.marginBottom="12px";const E=document.createElement("label");E.textContent=t(p,"File name"),E.style.display="block",E.style.marginBottom="6px";const f=document.createElement("input");f.type="text",f.value="Imported pad.pad",f.style.width="100%",f.style.boxSizing="border-box",f.style.marginBottom="12px";const y=document.createElement("p");y.style.color="var(--color-error, #c62828)",y.style.margin="0 0 12px 0",y.style.minHeight="20px";const w=document.createElement("button");w.type="button",w.className="primary",w.textContent=t(p,"Create");const g=v=>{i.remove(),n(v)};r.addEventListener("click",()=>g(null)),i.addEventListener("click",v=>{v.target===i&&g(null)}),s.addEventListener("blur",()=>{const v=s.value.trim();v.startsWith("http")&&(f.value=ot(v))});const N=()=>{const v=s.value.trim(),d=f.value.trim();if(v===""){y.textContent=t(p,"External pad URL is required.");return}if(d===""){y.textContent=t(p,"File name is required.");return}g({padUrl:v,name:d})};s.addEventListener("keydown",v=>{v.key==="Enter"&&(v.preventDefault(),N())}),f.addEventListener("keydown",v=>{v.key==="Enter"&&(v.preventDefault(),N())}),w.addEventListener("click",N),e.appendChild(a),e.appendChild(s),e.appendChild(E),e.appendChild(f),e.appendChild(y),e.appendChild(w),s.focus(),s.select()}),lt=()=>new Promise(n=>{const i=document.createElement("div");i.style.position="fixed",i.style.inset="0",i.style.background="rgba(0, 0, 0, 0.45)",i.style.display="flex",i.style.alignItems="center",i.style.justifyContent="center",i.style.zIndex="20000";const e=document.createElement("div");e.style.position="relative",e.style.background="var(--color-main-background, #fff)",e.style.borderRadius="10px",e.style.boxShadow="0 10px 30px rgba(0, 0, 0, 0.25)",e.style.padding="18px",e.style.width="min(420px, calc(100vw - 24px))";const r=document.createElement("h3");r.textContent=t(p,"Create public pad"),r.style.margin="0 26px 10px 0",r.style.fontSize="18px";const a=document.createElement("button");a.type="button",a.setAttribute("aria-label",t(p,"Close")),a.textContent="×",a.style.position="absolute",a.style.top="8px",a.style.right="10px",a.style.border="none",a.style.background="transparent",a.style.fontSize="22px",a.style.cursor="pointer",a.style.lineHeight="1";const s=document.createElement("p");s.textContent=t(p,"Choose how you want to create the public pad."),s.style.margin="0 0 14px 0",s.style.opacity="0.8";const E=document.createElement("div");E.style.display="grid",E.style.gap="8px";const f=document.createElement("button");f.type="button",f.className="primary",f.textContent=t(p,"Internal public pad");const y=document.createElement("button");y.type="button",y.textContent=t(p,"External public pad (URL)");const w=X(p,"etherpad-icon-color");w!==""&&(y.style.backgroundImage=`url(${w})`),y.style.backgroundRepeat="no-repeat",y.style.backgroundPosition="12px center",y.style.backgroundSize="16px 16px",y.style.paddingLeft="34px";const g=N=>{i.remove(),n(N)};a.addEventListener("click",()=>g(null)),f.addEventListener("click",()=>g("internal")),y.addEventListener("click",()=>g("external")),i.addEventListener("click",N=>{N.target===i&&g(null)}),E.appendChild(f),E.appendChild(y),e.appendChild(a),e.appendChild(r),e.appendChild(s),e.appendChild(E),i.appendChild(e),document.body.appendChild(i)}),F="data-epnc-single-file-share";let ce=0;const we=n=>{if(!n)return null;const i='a[href*="/s/'+n+'/download"],a[href*="/index.php/s/'+n+'/download"]',e=document.querySelector(i);return e instanceof HTMLAnchorElement?e:null},ct=()=>{const n=document.getElementById("initial-state-files_sharing-view");if(!(n instanceof HTMLInputElement))return"";const i=String(n.value||"").trim();if(i==="")return"";try{const e=window.atob(i);if(e.startsWith('"')||e.startsWith("{")||e.startsWith("[")){const r=JSON.parse(e);return typeof r=="string"?r:""}return e}catch{return""}},be=n=>{if(!n)return!1;const i=we(n);if(!i)return!1;let e;try{e=new URL(i.getAttribute("href")||"",window.location.origin)}catch{return!1}const r=(e.pathname||"").match(/(?:\/index\.php)?\/s\/([^/]+)\/download\/?$/);return!r||r[1]!==n?!1:!e.searchParams.has("files")},dt=()=>{const n=document.body;if(!(n instanceof HTMLElement))return{retry:!0};const i=U();if(!i)return n.removeAttribute(F),{retry:!1};const e=ct();return e==="public-file-share"?(n.setAttribute(F,"1"),{retry:!1}):e==="public-share"?(n.removeAttribute(F),{retry:!1}):we(i)?be(i)?(n.setAttribute(F,"1"),{retry:!1}):(n.removeAttribute(F),{retry:!1}):(n.removeAttribute(F),{retry:!0})},ut=()=>{const n=++ce,i=e=>{n===ce&&dt().retry&&e<30&&window.setTimeout(()=>i(e+1),180)};i(0)};(function(){const n="etherpad_nextcloud_open";let i=!1,e="";const r=Xe(),a=De(),s=o=>{const l=Number(o);return Number.isFinite(l)&&l>0?l:null},E=o=>{if(!o)return null;const l=[o.fileId,o.fileid,o.id,o.nodeId,o.nodeid,o.file_id];for(const m of l){const h=s(m);if(h!==null)return h}if(typeof o.get=="function")for(const m of["fileid","fileId","id"])try{const h=s(o.get(m));if(h!==null)return h}catch{}return null},f=o=>{if(!(o instanceof Element))return null;const l=[o.getAttribute("data-fileid"),o.getAttribute("data-file-id"),o.getAttribute("data-id"),o.getAttribute("data-node-id")];for(const h of l){const b=s(h);if(b!==null)return b}const m=o.matches("a[href]")?o:o.querySelector("a[href]");return m instanceof HTMLAnchorElement?ue(m.getAttribute("href")):null},y=o=>{if(!o||typeof o!="object")return null;const l=[o,o.fileInfoModel,o.fileInfo,o.model,o.node,o.file,o.attributes];for(const h of l){const b=E(h);if(b!==null)return b}const m=[o.fileElement,o.el,o.element,o.$file&&o.$file[0],o.$el&&o.$el[0],o.target,o.currentTarget];for(const h of m){const b=f(h);if(b!==null)return b}return null},w=async()=>{const o=await at();if(!o)return;const l=o.trim(),m=l.toLowerCase().endsWith(".pad")?l:l+".pad",h=R(K(),m);try{const b=await Me(h),M=b&&typeof b.file=="string"?b.file:h,O=b&&Number.isFinite(Number(b.file_id))?Number(b.file_id):null;await ee({path:M,fileId:O},{fallbackOpen:a,resolveOpenDir:Z})}catch(b){const M=b instanceof Error?b.message:t(p,"Could not create public pad.");window.alert(M)}},g=async()=>{const o=await st();if(!o)return;const l=o.padUrl.trim(),m=o.name.trim(),h=K(),b=m.toLowerCase().endsWith(".pad")?m:m+".pad",M=R(h,b);try{const O=await Ie(M,l),W=O&&typeof O.file=="string"?O.file:M,he=O&&Number.isFinite(Number(O.file_id))?Number(O.file_id):null;await ee({path:W,fileId:he},{fallbackOpen:a,resolveOpenDir:Z})}catch(O){const W=O instanceof Error?O.message:t(p,"Could not import public pad URL.");window.alert(W)}},N=async()=>{const o=await lt();if(o==="internal"){await w();return}o==="external"&&await g()},v=(o,l)=>{if(!(o instanceof HTMLAnchorElement)||!l)return null;const m=o.getAttribute("href")||"",h=Ne(m);if(h&&h.token===l&&B(h.path))return h.path;const b=Le(m);return B(b)?b:null},d=()=>{if(window.OCA&&window.OCA.EtherpadNextcloudClickInterceptorRegistered===!0)return;window.OCA&&(window.OCA.EtherpadNextcloudClickInterceptorRegistered=!0);const o=async l=>{if(!l||l.defaultPrevented)return;const m=U();if(!m||l.type==="click"&&(l.button!==0||l.metaKey||l.ctrlKey||l.shiftKey||l.altKey)||!(l.target instanceof Element))return;const h=l.target.closest("a[href]");if(!(h instanceof HTMLAnchorElement))return;const b=v(h,m);if(b){if(l.preventDefault(),l.stopPropagation(),typeof l.stopImmediatePropagation=="function"&&l.stopImmediatePropagation(),q()){await a({path:b,fileId:null});return}window.location.assign(D(m,b))}};document.addEventListener("click",o,!0)},u=()=>{window.OCA&&window.OCA.Files&&window.OCA.Files.fileActions&&(window.OCA.Files.fileActions.register(Y,n,xe(),X(p,"etherpad-icon-color"),(o,l)=>{const m=l&&l.dir||K(),h=R(m,o),b=y(l);a({path:h,fileId:b})},t(p,"Open in Etherpad")),window.OCA.Files.fileActions.setDefault(Y,n))},c=rt({isFilesAppRoute:k,onCreatePublicPad:N}),C=async()=>{if(q())return;const o=z();if(!o)return;const l=new URLSearchParams(window.location.search||"");if(!(l.get("epNoRedirect")==="1"||l.get("details")==="1"||l.get("opendetails")==="true")&&!(window.location.pathname||"").includes("/apps/"+p+"/"))try{const m=await $(o);m&&m.is_pad&&m.viewer_url&&window.location.assign(m.viewer_url)}catch{}},A=async()=>{if(q()||!k())return;const o=z();if(!o)return;const l=new URLSearchParams(window.location.search||"");if(!(l.get("openfile")==="true"||l.get("opendetails")==="true"||l.get("details")==="1"))try{const m=await $(o);if(!m||!m.is_pad)return;const h=l.get("dir")||"/",b=T("/apps/files/files")+"?dir="+encodeURIComponent(h);window.location.href!==b&&window.location.assign(b)}catch{}},x=()=>{const o=U();if(!o)return;const l=new URLSearchParams(window.location.search||""),m=l.get("files")||"";if(q()){if(!B(m))return;const M=l.get("path")||"/",O=R(M,m);a({path:O,fileId:null});return}if(!B(m)){if((window.location.pathname||"").includes("/apps/"+p+"/public/"))return;be(o)&&window.location.assign(D(o,""));return}const h=l.get("path")||"/",b=R(h,m);window.location.assign(D(o,b))},S=()=>{const o=(window.location.pathname||"")+"?"+(window.location.search||"");o!==e&&(e=o,k()&&c(),ut(),A(),C(),x(),r.scheduleRefresh(0))},L=()=>{const o="EtherpadNextcloudRouteWatchInstalled";if(window.OCA&&window.OCA[o]===!0)return;window.OCA&&(window.OCA[o]=!0);const l=()=>window.setTimeout(S,0);window.addEventListener("popstate",l),window.addEventListener("hashchange",l);const m=h=>{const b=window.history&&window.history[h];typeof b=="function"&&(window.history[h]=function(...M){const O=b.apply(this,M);return l(),O})};m("pushState"),m("replaceState")},I=()=>{i||(i=!0,L(),S(),u(),d(),r.installObserver(),r.scheduleRefresh(200))};document.readyState==="loading"?document.addEventListener("DOMContentLoaded",I,{once:!0}):I()})();
+import{a as P,A as y,o as ae,b as ye,c as we,d as G,n as be,e as he,M as J,f as ge}from"./oc-compat-Cn8-bqPr.chunk.mjs";import{p as q,a as R,v as H,f as U,b as ve,r as se,n as F,g as W,c as le,i as _,d as Ce,e as Ee}from"./urls-D-89WmZ3.chunk.mjs";const M=new Map,z=async e=>{const i=String(e);if(M.has(i))return M.get(i);const n=P("/apps/"+y+"/api/v1/pads/resolve")+"?fileId="+encodeURIComponent(i),r=I(n,{method:"GET",headers:{Accept:"application/json"}},"Pad resolve failed.").catch(a=>{throw M.delete(i),a});return M.set(i,r),r},xe=async e=>{const i="path:"+String(e);if(M.has(i))return M.get(i);const n=P("/apps/"+y+"/api/v1/pads/resolve")+"?file="+encodeURIComponent(e),r=I(n,{method:"GET",headers:{Accept:"application/json"}},"Pad resolve by path failed.").catch(a=>{throw M.delete(i),a});return M.set(i,r),r},D=async e=>{const i=P("/apps/"+y+"/api/v1/pads/sync-status/"+encodeURIComponent(String(e)));return I(i,{method:"GET",headers:{Accept:"application/json"}},"Sync status check failed.")},Ae=async(e,i)=>{let n=P("/apps/"+y+"/api/v1/pads/sync/"+encodeURIComponent(String(e)));return n+="?force=1",I(n,{method:"POST",headers:{Accept:"application/json",requesttoken:ae()}},"Pad sync failed.")},Ne=async(e,i)=>{const n=P("/apps/"+y+"/api/v1/pads/from-url"),r=new URLSearchParams;return r.set("file",e),r.set("padUrl",i),I(n,{method:"POST",headers:ce(),body:r.toString()},"Could not import public pad URL.")},Se=async e=>{const i=P("/apps/"+y+"/api/v1/pads"),n=new URLSearchParams;return n.set("file",e),n.set("accessMode","public"),I(i,{method:"POST",headers:ce(),body:n.toString()},"Could not create public pad.")},ce=()=>({Accept:"application/json","Content-Type":"application/x-www-form-urlencoded;charset=UTF-8",requesttoken:ae()}),I=async(e,i,n)=>{const r=await fetch(e,{...i,credentials:"same-origin"}),a=await r.json().catch(()=>({}));if(!r.ok)throw new Error(a&&a.message||n);return a},k=()=>!!(window.OCA&&window.OCA.Viewer&&typeof window.OCA.Viewer.open=="function"),L=()=>(window.location.pathname||"").includes("/apps/files"),Le=800,Me=180,Pe=(e,i)=>{const n=window.OCP&&window.OCP.Files&&window.OCP.Files.Router;return!n||typeof n.goToRoute!="function"||!e||!Number.isFinite(e)?!1:(n.goToRoute(null,{view:"files",fileid:String(e)},{dir:se(i),editing:"false",openfile:"true"}),window.setTimeout(()=>{if((window.location.pathname||"").includes("/apps/files/files/"+String(e)))return;const r=U(e,i);window.location.assign(r)},Me),!0)},Oe=()=>{let e=null,i=0;return async n=>{const r=String(n.fileId??"")+"|"+String(n.path??""),a=Date.now();if(e===r&&a-i<Le)return;e=r,i=a;const s=q(),b=s!==null&&s!=="",d=()=>{if(b){window.location.assign(H(s,n.path||""));return}if(L()&&n.fileId!==null&&n.fileId!==void 0&&Number.isFinite(Number(n.fileId))){const N=n.path||"/";window.location.assign(U(Number(n.fileId),N));return}const h=L()?R():null;if((n.fileId===null||n.fileId===void 0)&&h){const N=n.path||"/";window.location.assign(U(h,N));return}if(n.fileId!==null&&n.fileId!==void 0&&Number.isFinite(Number(n.fileId))){const N=n.path||"/";window.location.assign(U(Number(n.fileId),N));return}n.path&&window.location.assign(ve(n.path))};if(!k()){d();return}let f=n.path||"",w=n.fileId??null;if(!f&&n.fileId!==null&&n.fileId!==void 0)try{const h=await z(n.fileId);f=h&&typeof h.path=="string"?h.path:"",w=h&&Number.isFinite(Number(h.file_id))?Number(h.file_id):w}catch{f=""}if(!f){d();return}if((!w||!Number.isFinite(Number(w)))&&f&&!b)try{const h=await xe(f);w=h&&Number.isFinite(Number(h.file_id))?Number(h.file_id):w}catch{}if((!w||!Number.isFinite(Number(w)))&&L()){const h=R();h&&(w=h)}if(L()){if(w&&Number.isFinite(Number(w))&&Pe(Number(w),f))return;d();return}try{window.OCA.Viewer.open({path:f})}catch{d()}}},ke=120,Ie=900,Te=async(e,i={})=>{if(k()&&e.path)try{const n=Number(e.fileId);L()&&Number.isFinite(n)&&n>0&&(_e(n,e.path,i.resolveOpenDir),await Fe()),await qe(e.path),window.OCA.Viewer.open({path:e.path,onClose:Ue});return}catch{}typeof i.fallbackOpen=="function"&&await i.fallbackOpen(e)},Fe=()=>new Promise(e=>{window.setTimeout(e,ke)}),qe=async e=>{const i=ye(e);i!==""&&(we("editor:file:created",i),await Re())},Re=()=>new Promise(e=>{window.setTimeout(e,Ie)}),_e=(e,i,n)=>{const r=window.OCP&&window.OCP.Files&&window.OCP.Files.Router;!r||typeof r.goToRoute!="function"||r.goToRoute(null,{...r.params||{},view:r.params&&r.params.view?r.params.view:"files",fileid:String(e)},{...r.query||{},dir:He(i,n),editing:"false",openfile:"true"},!0)},He=(e,i)=>{if(typeof i=="function")return i(e);const n=String(e||""),r=n.lastIndexOf("/");return r>0?n.substring(0,r):"/"},Ue=()=>{const e=window.OCP&&window.OCP.Files&&window.OCP.Files.Router;if(!e||typeof e.goToRoute!="function")return;const i={...e.query||{}};delete i.openfile,delete i.editing,e.goToRoute(null,e.params||{},i)},ze=e=>{try{const i=new URL(e,window.location.origin),n=decodeURIComponent(i.pathname||"").match(/\/p\/([^/]+)$/);if(n&&n[1]){const r=n[1].replace(/[^a-zA-Z0-9._-]+/g,"-").replace(/^-+|-+$/g,"");if(r!=="")return r+".pad"}}catch{}return"Imported pad.pad"},de=e=>{const i=document.createElement("div");i.style.position="fixed",i.style.inset="0",i.style.background="rgba(0, 0, 0, 0.45)",i.style.display="flex",i.style.alignItems="center",i.style.justifyContent="center",i.style.zIndex="20000";const n=document.createElement("div");n.style.position="relative",n.style.background="var(--color-main-background, #fff)",n.style.borderRadius="10px",n.style.boxShadow="0 10px 30px rgba(0, 0, 0, 0.25)",n.style.padding="18px",n.style.width="min(460px, calc(100vw - 24px))";const r=document.createElement("button");r.type="button",r.setAttribute("aria-label",t(y,"Close")),r.textContent="×",r.style.position="absolute",r.style.top="8px",r.style.right="10px",r.style.border="none",r.style.background="transparent",r.style.fontSize="22px",r.style.cursor="pointer",r.style.lineHeight="1";const a=document.createElement("h3");return a.textContent=e,a.style.margin="0 26px 10px 0",a.style.fontSize="18px",n.appendChild(r),n.appendChild(a),i.appendChild(n),document.body.appendChild(i),{overlay:i,dialog:n,closeButton:r}},Be=()=>new Promise(e=>{const{overlay:i,dialog:n,closeButton:r}=de(t(y,"Internal public pad")),a=document.createElement("label");a.textContent=t(y,"File name"),a.style.display="block",a.style.marginBottom="6px";const s=document.createElement("input");s.type="text",s.value=t(y,"Public pad")+".pad",s.style.width="100%",s.style.boxSizing="border-box",s.style.marginBottom="12px";const b=document.createElement("p");b.style.color="var(--color-error, #c62828)",b.style.margin="0 0 12px 0",b.style.minHeight="20px";const d=document.createElement("button");d.type="button",d.className="primary",d.textContent=t(y,"Create");const f=w=>{i.remove(),e(w)};r.addEventListener("click",()=>f(null)),i.addEventListener("click",w=>{w.target===i&&f(null)}),s.addEventListener("keydown",w=>{w.key==="Enter"&&(w.preventDefault(),d.click())}),d.addEventListener("click",()=>{const w=s.value.trim();if(w===""){b.textContent=t(y,"File name is required.");return}f(w)}),n.appendChild(a),n.appendChild(s),n.appendChild(b),n.appendChild(d),s.focus(),s.select()}),je=()=>new Promise(e=>{const{overlay:i,dialog:n,closeButton:r}=de(t(y,"External public pad (URL)")),a=document.createElement("label");a.textContent=t(y,"External pad URL"),a.style.display="block",a.style.marginBottom="6px";const s=document.createElement("input");s.type="url",s.value="https://",s.placeholder="https://",s.style.width="100%",s.style.boxSizing="border-box",s.style.marginBottom="12px";const b=document.createElement("label");b.textContent=t(y,"File name"),b.style.display="block",b.style.marginBottom="6px";const d=document.createElement("input");d.type="text",d.value="Imported pad.pad",d.style.width="100%",d.style.boxSizing="border-box",d.style.marginBottom="12px";const f=document.createElement("p");f.style.color="var(--color-error, #c62828)",f.style.margin="0 0 12px 0",f.style.minHeight="20px";const w=document.createElement("button");w.type="button",w.className="primary",w.textContent=t(y,"Create");const h=v=>{i.remove(),e(v)};r.addEventListener("click",()=>h(null)),i.addEventListener("click",v=>{v.target===i&&h(null)}),s.addEventListener("blur",()=>{const v=s.value.trim();v.startsWith("http")&&(d.value=ze(v))});const N=()=>{const v=s.value.trim(),u=d.value.trim();if(v===""){f.textContent=t(y,"External pad URL is required.");return}if(u===""){f.textContent=t(y,"File name is required.");return}h({padUrl:v,name:u})};s.addEventListener("keydown",v=>{v.key==="Enter"&&(v.preventDefault(),N())}),d.addEventListener("keydown",v=>{v.key==="Enter"&&(v.preventDefault(),N())}),w.addEventListener("click",N),n.appendChild(a),n.appendChild(s),n.appendChild(b),n.appendChild(d),n.appendChild(f),n.appendChild(w),s.focus(),s.select()}),De=()=>new Promise(e=>{const i=document.createElement("div");i.style.position="fixed",i.style.inset="0",i.style.background="rgba(0, 0, 0, 0.45)",i.style.display="flex",i.style.alignItems="center",i.style.justifyContent="center",i.style.zIndex="20000";const n=document.createElement("div");n.style.position="relative",n.style.background="var(--color-main-background, #fff)",n.style.borderRadius="10px",n.style.boxShadow="0 10px 30px rgba(0, 0, 0, 0.25)",n.style.padding="18px",n.style.width="min(420px, calc(100vw - 24px))";const r=document.createElement("h3");r.textContent=t(y,"Create public pad"),r.style.margin="0 26px 10px 0",r.style.fontSize="18px";const a=document.createElement("button");a.type="button",a.setAttribute("aria-label",t(y,"Close")),a.textContent="×",a.style.position="absolute",a.style.top="8px",a.style.right="10px",a.style.border="none",a.style.background="transparent",a.style.fontSize="22px",a.style.cursor="pointer",a.style.lineHeight="1";const s=document.createElement("p");s.textContent=t(y,"Choose how you want to create the public pad."),s.style.margin="0 0 14px 0",s.style.opacity="0.8";const b=document.createElement("div");b.style.display="grid",b.style.gap="8px";const d=document.createElement("button");d.type="button",d.className="primary",d.textContent=t(y,"Internal public pad");const f=document.createElement("button");f.type="button",f.textContent=t(y,"External public pad (URL)");const w=G(y,"etherpad-icon-color");w!==""&&(f.style.backgroundImage=`url(${w})`),f.style.backgroundRepeat="no-repeat",f.style.backgroundPosition="12px center",f.style.backgroundSize="16px 16px",f.style.paddingLeft="34px";const h=N=>{i.remove(),e(N)};a.addEventListener("click",()=>h(null)),d.addEventListener("click",()=>h("internal")),f.addEventListener("click",()=>h("external")),i.addEventListener("click",N=>{N.target===i&&h(null)}),b.appendChild(d),b.appendChild(f),n.appendChild(a),n.appendChild(r),n.appendChild(s),n.appendChild(b),i.appendChild(n),document.body.appendChild(i)}),Z=e=>e.toLowerCase().endsWith(".pad")?e:e+".pad",$e=(e,i)=>({path:e&&typeof e.file=="string"?e.file:i,fileId:e&&Number.isFinite(Number(e.file_id))?Number(e.file_id):null}),Ve=({openPadInNativeViewer:e})=>{const i=(a,s)=>Te($e(a,s),{fallbackOpen:e,resolveOpenDir:se}),n=async()=>{const a=await Be();if(!a)return;const s=a.trim(),b=F(W(),Z(s));try{const d=await Se(b);await i(d,b)}catch(d){const f=d instanceof Error?d.message:t(y,"Could not create public pad.");window.alert(f)}},r=async()=>{const a=await je();if(!a)return;const s=a.padUrl.trim(),b=a.name.trim(),d=F(W(),Z(b));try{const f=await Ne(d,s);await i(f,d)}catch(f){const w=f instanceof Error?f.message:t(y,"Could not import public pad URL.");window.alert(w)}};return async()=>{const a=await De();if(a==="internal"){await n();return}a==="external"&&await r()}},$="data-epnc-sidebar-sync-panel",K="data-epnc-sidebar-sync-mount",ue=8e3,We=6e4,pe=15e3,Ke=8e3,Ge=()=>{const e=document.documentElement,i=document.body,n=(e?.className||"")+" "+(i?.className||"");if(/theme[-_]?dark|dark-theme|theme--dark/i.test(n))return!0;const r=window.matchMedia&&window.matchMedia("(prefers-color-scheme: dark)");return!!(r&&r.matches)},Q=e=>{if(!(e instanceof HTMLElement))return!1;const i=window.getComputedStyle(e);return i.display==="none"||i.visibility==="hidden"?!1:i.position==="fixed"?!0:e.offsetParent!==null},Je=e=>{const i=Number(e);return Number.isFinite(i)&&i>0?i:null},Ze=()=>{const e=R();if(e!==null)return e;const i=new URLSearchParams(window.location.search||""),n=Number(i.get("fileid")||"");return Number.isFinite(n)&&n>0?n:null},X=()=>{const e=["[data-cy-files-sidebar]","#app-sidebar-vue .app-sidebar","#app-sidebar-vue",".app-sidebar"];for(const i of e){const n=document.querySelector(i);if(n instanceof HTMLElement)return n}return null},Qe=e=>{if(!(e instanceof HTMLElement))return null;let i=e.querySelector("["+K+'="1"]');i instanceof HTMLElement||(i=document.createElement("div"),i.className="epnc-sidebar-sync-mount",i.setAttribute(K,"1"));const n=e.querySelector("[data-cy-files-sidebar-tabs], .app-sidebar-tabs");if(n&&n.parentNode instanceof HTMLElement){const r=n.parentNode;return(i.parentNode!==r||i.nextSibling!==n)&&r.insertBefore(i,n),i}return i.parentNode!==e&&e.insertBefore(i,e.firstChild),i},Y=e=>{const i=Ze();if(i!==null)return i;if(!(e instanceof HTMLElement))return null;const n=["[data-fileid]","[data-file-id]","[data-node-id]","[data-id]"];for(const a of n){const s=e.querySelector(a);if(!(s instanceof HTMLElement))continue;const b=[s.getAttribute("data-fileid"),s.getAttribute("data-file-id"),s.getAttribute("data-node-id"),s.getAttribute("data-id"),s.dataset&&(s.dataset.fileid||s.dataset.fileId)];for(const d of b){const f=Je(d);if(f!==null)return f}}const r=e.querySelectorAll("a[href]");for(const a of r){if(!(a instanceof HTMLAnchorElement))continue;const s=le(a.getAttribute("href"));if(s!==null)return s}return null},Xe=e=>e==="synced"?"etherpad-icon-color":Ge()?"etherpad-icon-white":"etherpad-icon-black",Ye=e=>e==="synced"?t(y,"Synchronized"):e==="syncing"?t(y,"Saving..."):e==="error"?t(y,"Sync failed"):e==="unavailable"?t(y,"Sync status unavailable"):e==="loading"?t(y,"Checking sync status..."):t(y,"Sync pending"),V=e=>e==="synced"?"synced":e==="out_of_sync"?"pending":e==="unavailable"?"unavailable":"error",et=(e,i)=>e==="synced"?Math.min(We,i+Ke):e==="pending"?ue:pe,tt=()=>{let e=null,i=null,n=null,r=0;const a=()=>{n!==null&&(window.clearTimeout(n),n=null)},s=()=>{const u=document.querySelector("["+$+'="1"]');u&&u.parentNode&&u.parentNode.removeChild(u);const m=document.querySelector("["+K+'="1"]');m&&m.parentNode&&m.parentNode.removeChild(m),a()},b=(u,m)=>{if(!(u instanceof HTMLElement))return;const c=String(m||"").trim();u.setAttribute("data-public-open-url",c);const C=u.querySelector("[data-epnc-sidebar-open-public-button]");if(!(C instanceof HTMLButtonElement))return;const A=c!=="";C.style.display=A?"inline-flex":"none",C.disabled=!A},d=(u,m,c,C,A=!0)=>{if(!(u instanceof HTMLElement))return;const x=u.querySelector("[data-epnc-sidebar-sync-icon]"),S=u.querySelector("[data-epnc-sidebar-sync-label]"),o=u.querySelector("[data-epnc-sidebar-sync-button]");if(u.setAttribute("data-file-id",String(c)),u.setAttribute("data-sync-state",m),u.setAttribute("data-sync-in-flight",C?"1":"0"),S instanceof HTMLElement&&(S.textContent=Ye(m)),x instanceof HTMLElement){const l=Xe(m),p=G(y,l);p!==""&&(x.style.backgroundImage=`url(${p})`)}o instanceof HTMLButtonElement&&(o.disabled=C||!A||c<=0,o.setAttribute("aria-busy",C?"true":"false"))},f=(u,m)=>{if(!(m instanceof HTMLElement))return null;let c=m.querySelector("["+$+'="1"]');if(!(c instanceof HTMLElement)){c=document.createElement("section"),c.className="epnc-sidebar-sync-panel",c.setAttribute($,"1"),c.setAttribute("data-file-id",String(u));const C=document.createElement("div");C.className="epnc-sidebar-sync-panel__header";const A=document.createElement("span");A.className="epnc-sidebar-sync-panel__icon",A.setAttribute("data-epnc-sidebar-sync-icon","1"),A.setAttribute("aria-hidden","true");const x=document.createElement("span");x.className="epnc-sidebar-sync-panel__status",x.setAttribute("data-epnc-sidebar-sync-label","1"),C.appendChild(A),C.appendChild(x);const S=document.createElement("button");S.type="button",S.className="button-vue button-vue--size-normal button-vue--vue-secondary epnc-sidebar-sync-panel__button",S.setAttribute("data-epnc-sidebar-sync-button","1"),S.textContent=t(y,"Pad in Datei speichern"),S.addEventListener("click",async l=>{l.preventDefault(),l.stopPropagation();const p=Number(c.getAttribute("data-file-id")||"");if(!(!Number.isFinite(p)||p<=0)){d(c,"syncing",p,!0,!0);try{await Ae(p,!0);const g=await D(p),E=V(g&&g.status);d(c,E,p,!1,!0)}catch{d(c,"error",p,!1,!0)}}});const o=document.createElement("button");o.type="button",o.className="button-vue button-vue--size-normal button-vue--vue-tertiary epnc-sidebar-sync-panel__button epnc-sidebar-sync-panel__button--public-open",o.setAttribute("data-epnc-sidebar-open-public-button","1"),o.textContent=t(y,"Pad in neuem Tab öffnen"),o.style.display="none",o.addEventListener("click",l=>{l.preventDefault(),l.stopPropagation();const p=String(c.getAttribute("data-public-open-url")||"").trim();p!==""&&window.open(p,"_blank","noopener,noreferrer")}),c.appendChild(C),c.appendChild(S),c.appendChild(o)}return c.parentNode!==m&&m.appendChild(c),c},w=(u,m)=>{if(a(),!(u instanceof HTMLElement)||!Number.isFinite(m)||m<=0)return;let c=ue;const C=x=>{n=window.setTimeout(()=>{A()},Math.max(1e3,x))},A=async()=>{if(!(u instanceof HTMLElement)||!document.body.contains(u)){a();return}const x=X();if(!(x instanceof HTMLElement)||!Q(x)||!Q(u)){a();return}if(Y(x)!==m){a();return}if(document.visibilityState!=="visible"||u.getAttribute("data-sync-in-flight")==="1"){C(c);return}try{const S=await D(m),o=V(S&&S.status);d(u,o,m,!1,!0),c=et(o,c)}catch{d(u,"error",m,!1,!0),c=pe}C(c)};C(c)},h=()=>{i!==null&&(window.clearTimeout(i),i=null)},N=async()=>{const u=++r;if(!L()||q()){s();return}const m=X();if(!(m instanceof HTMLElement)){s();return}const c=Y(m);if(!c){s();return}const C=Qe(m);if(!(C instanceof HTMLElement)){s();return}let A;try{A=await z(c)}catch{s();return}if(u!==r)return;if(!A||A.is_pad!==!0){s();return}const x=f(c,C);if(!(x instanceof HTMLElement)){s();return}const S=A&&typeof A.public_open_url=="string"?A.public_open_url.trim():"";b(x,S),d(x,"loading",c,!1,!0);try{const o=await D(c);if(u!==r)return;const l=V(o&&o.status);d(x,l,c,!1,!0),w(x,c)}catch{if(u!==r)return;d(x,"error",c,!1,!0),w(x,c)}},v=(u=0)=>{h(),i=window.setTimeout(()=>{i=null,N()},Math.max(0,u))};return{installObserver:()=>{if(e!==null||!L())return;const u=new MutationObserver(m=>{for(const c of m)for(const C of c.addedNodes)if(C instanceof Element&&(C.matches("[data-cy-files-sidebar], #app-sidebar-vue, .app-sidebar")||C.querySelector("[data-cy-files-sidebar], #app-sidebar-vue, .app-sidebar"))){v(60);return}});u.observe(document.body,{childList:!0,subtree:!0}),e=u},scheduleRefresh:v}},ee=y+"_public_pad",te="icon-filetype-etherpad-nextcloud-pad",ne='<svg xmlns="http://www.w3.org/2000/svg" xml:space="preserve" fill-rule="evenodd" stroke-linejoin="round" stroke-miterlimit="2" clip-rule="evenodd" viewBox="0 0 355 355"><path fill="#5ac395" d="M317 89v177c0 28-23 51-51 51H89c-28 0-51-23-51-51V89c0-28 23-51 51-51h177c28 0 51 23 51 51"/><path fill="#4aa57f" fill-rule="nonzero" d="M240 144q-4-3-8 0-6 6 0 9a36 36 0 0 1 0 52 6 6 0 0 0 3 12q3 0 5-3c19-18 19-50 0-70"/><path fill="#479a76" fill-rule="nonzero" d="M267 122q-4-5-9 0-4 4 0 10c25 26 25 68 0 93q-4 5 0 9 5 4 9 0c30-31 30-81 0-112"/><path fill="#fff" d="M192 130v2q-1 8-10 10H76q-9-2-11-10v-2q1-10 11-10h106q9 0 10 10m24 51v2q-1 11-12 11H78q-10 0-12-11v-2q2-10 12-11h126q12 1 12 11"/><path fill="#fff" d="M216 181v2q-1 11-12 11H78q-10 0-12-11v-2q2-10 12-11h126q12 1 12 11m-57 52v1q-1 11-10 11H76q-9 0-11-11v-1q1-10 11-11h73q9 1 10 11"/><path fill="#fff" d="M159 233v1q-1 11-10 11H76q-9 0-11-11v-1q1-10 11-11h73q9 1 10 11"/></svg>',ie=98,nt=120,it=500,re=()=>{const e=be();return e!==null&&e>=33},T=e=>(e&&e.message?String(e.message):"").toLowerCase().includes("duplicate"),rt=e=>{if(!e||typeof e!="object")return null;if(e.hasCreatePermission===!1)return!1;if(e.hasCreatePermission===!0)return!0;if(typeof e.canCreate=="boolean")return e.canCreate;const i=he(),n=[e.permissions,e.permission,e.attributes&&e.attributes.permissions,typeof e.get=="function"?e.get("permissions"):null];for(const r of n){const a=Number(r);if(Number.isFinite(a))return(a&i)===i}return null},ot=(...e)=>{for(const i of e){const n=rt(i);if(n!==null)return n}return!0},at=()=>{const e=[window.OCP&&window.OCP.Files,window.OCA&&window.OCA.Files];for(const n of e)if(!(!n||typeof n!="object")){if(typeof n.addNewFileMenuEntry=="function")return{register:n.addNewFileMenuEntry.bind(n),unregister:typeof n.removeNewFileMenuEntry=="function"?n.removeNewFileMenuEntry.bind(n):null};if(typeof n.getNewFileMenu=="function"){let r=null;try{r=n.getNewFileMenu()}catch{r=null}if(r&&typeof r.registerEntry=="function")return{register:r.registerEntry.bind(r),unregister:typeof r.unregisterEntry=="function"?r.unregisterEntry.bind(r):null}}}const i=window._nc_newfilemenu;return i&&typeof i=="object"&&typeof i.registerEntry=="function"?{register:i.registerEntry.bind(i),unregister:typeof i.unregisterEntry=="function"?i.unregisterEntry.bind(i):null}:null},st=({isFilesAppRoute:e,onCreatePublicPad:i})=>{let n=!1,r=!1,a=!1,s=0;const b=()=>{i()},d=(...v)=>ot(...v),f=()=>({id:ee,displayName:t(y,"Public pad"),...re()?{iconSvgInline:ne}:{iconClass:te},order:ie,enabled:d,handler:b}),w=()=>({id:ee,displayName:t(y,"Public pad"),templateName:t(y,"Public pad"),iconClass:te,...re()?{iconSvgInline:ne}:{},fileType:"file",order:ie,actionHandler:b}),h=async()=>{const v=at();if(!v)return!1;const u=f();try{return v.register(u),n=!0,!0}catch(m){return T(m)?(n=!0,!0):!1}},N=()=>{if(r)return!0;const v=w(),u=()=>{const m=[window.OCA&&window.OCA.Files&&window.OCA.Files.NewFileMenu,window.OCA&&window.OCA.Files&&window.OCA.Files.newFileMenu];for(const c of m)if(!(!c||typeof c!="object")){if(typeof c.addMenuEntry=="function")try{return c.addMenuEntry(v),!0}catch(C){if(T(C))return!0}if(typeof c.registerEntry=="function")try{return c.registerEntry(f()),!0}catch(C){if(T(C))return!0}}return!1};try{return u()?(r=!0,!0):(!a&&window.OC&&window.OC.Plugins&&typeof window.OC.Plugins.register=="function"&&(window.OC.Plugins.register("OCA.Files.NewFileMenu",{attach(m){if(!(!m||typeof m.addMenuEntry!="function"))try{m.addMenuEntry(v),r=!0}catch(c){T(c)&&(r=!0)}}}),a=!0),r)}catch(m){return T(m)?(r=!0,!0):!1}};return()=>{if(!e()||n||r)return;const v=++s,u=async m=>{v!==s||n||r||await h()||N()||m<nt&&window.setTimeout(()=>{u(m+1)},it)};u(0)}},O="data-epnc-single-file-share";let oe=0;const fe=e=>{if(!e)return null;const i='a[href*="/s/'+e+'/download"],a[href*="/index.php/s/'+e+'/download"]',n=document.querySelector(i);return n instanceof HTMLAnchorElement?n:null},lt=()=>{const e=document.getElementById("initial-state-files_sharing-view");if(!(e instanceof HTMLInputElement))return"";const i=String(e.value||"").trim();if(i==="")return"";try{const n=window.atob(i);if(n.startsWith('"')||n.startsWith("{")||n.startsWith("[")){const r=JSON.parse(n);return typeof r=="string"?r:""}return n}catch{return""}},me=e=>{if(!e)return!1;const i=fe(e);if(!i)return!1;let n;try{n=new URL(i.getAttribute("href")||"",window.location.origin)}catch{return!1}const r=(n.pathname||"").match(/(?:\/index\.php)?\/s\/([^/]+)\/download\/?$/);return!r||r[1]!==e?!1:!n.searchParams.has("files")},ct=()=>{const e=document.body;if(!(e instanceof HTMLElement))return{retry:!0};const i=q();if(!i)return e.removeAttribute(O),{retry:!1};const n=lt();return n==="public-file-share"?(e.setAttribute(O,"1"),{retry:!1}):n==="public-share"?(e.removeAttribute(O),{retry:!1}):fe(i)?me(i)?(e.setAttribute(O,"1"),{retry:!1}):(e.removeAttribute(O),{retry:!1}):(e.removeAttribute(O),{retry:!0})},dt=()=>{const e=++oe,i=n=>{e===oe&&ct().retry&&n<30&&window.setTimeout(()=>i(n+1),180)};i(0)};(function(){const e="etherpad_nextcloud_open";let i=!1,n="";const r=tt(),a=Oe(),s=Ve({openPadInNativeViewer:a}),b=o=>{const l=Number(o);return Number.isFinite(l)&&l>0?l:null},d=o=>{if(!o)return null;const l=[o.fileId,o.fileid,o.id,o.nodeId,o.nodeid,o.file_id];for(const p of l){const g=b(p);if(g!==null)return g}if(typeof o.get=="function")for(const p of["fileid","fileId","id"])try{const g=b(o.get(p));if(g!==null)return g}catch{}return null},f=o=>{if(!(o instanceof Element))return null;const l=[o.getAttribute("data-fileid"),o.getAttribute("data-file-id"),o.getAttribute("data-id"),o.getAttribute("data-node-id")];for(const g of l){const E=b(g);if(E!==null)return E}const p=o.matches("a[href]")?o:o.querySelector("a[href]");return p instanceof HTMLAnchorElement?le(p.getAttribute("href")):null},w=o=>{if(!o||typeof o!="object")return null;const l=[o,o.fileInfoModel,o.fileInfo,o.model,o.node,o.file,o.attributes];for(const g of l){const E=d(g);if(E!==null)return E}const p=[o.fileElement,o.el,o.element,o.$file&&o.$file[0],o.$el&&o.$el[0],o.target,o.currentTarget];for(const g of p){const E=f(g);if(E!==null)return E}return null},h=(o,l)=>{if(!(o instanceof HTMLAnchorElement)||!l)return null;const p=o.getAttribute("href")||"",g=Ce(p);if(g&&g.token===l&&_(g.path))return g.path;const E=Ee(p);return _(E)?E:null},N=()=>{if(window.OCA&&window.OCA.EtherpadNextcloudClickInterceptorRegistered===!0)return;window.OCA&&(window.OCA.EtherpadNextcloudClickInterceptorRegistered=!0);const o=async l=>{if(!l||l.defaultPrevented)return;const p=q();if(!p||l.type==="click"&&(l.button!==0||l.metaKey||l.ctrlKey||l.shiftKey||l.altKey)||!(l.target instanceof Element))return;const g=l.target.closest("a[href]");if(!(g instanceof HTMLAnchorElement))return;const E=h(g,p);if(E){if(l.preventDefault(),l.stopPropagation(),typeof l.stopImmediatePropagation=="function"&&l.stopImmediatePropagation(),k()){await a({path:E,fileId:null});return}window.location.assign(H(p,E))}};document.addEventListener("click",o,!0)},v=()=>{window.OCA&&window.OCA.Files&&window.OCA.Files.fileActions&&(window.OCA.Files.fileActions.register(J,e,ge(),G(y,"etherpad-icon-color"),(o,l)=>{const p=l&&l.dir||W(),g=F(p,o),E=w(l);a({path:g,fileId:E})},t(y,"Open in Etherpad")),window.OCA.Files.fileActions.setDefault(J,e))},u=st({isFilesAppRoute:L,onCreatePublicPad:s}),m=async()=>{if(k())return;const o=R();if(!o)return;const l=new URLSearchParams(window.location.search||"");if(!(l.get("epNoRedirect")==="1"||l.get("details")==="1"||l.get("opendetails")==="true")&&!(window.location.pathname||"").includes("/apps/"+y+"/"))try{const p=await z(o);p&&p.is_pad&&p.viewer_url&&window.location.assign(p.viewer_url)}catch{}},c=async()=>{if(k()||!L())return;const o=R();if(!o)return;const l=new URLSearchParams(window.location.search||"");if(!(l.get("openfile")==="true"||l.get("opendetails")==="true"||l.get("details")==="1"))try{const p=await z(o);if(!p||!p.is_pad)return;const g=l.get("dir")||"/",E=P("/apps/files/files")+"?dir="+encodeURIComponent(g);window.location.href!==E&&window.location.assign(E)}catch{}},C=()=>{const o=q();if(!o)return;const l=new URLSearchParams(window.location.search||""),p=l.get("files")||"";if(k()){if(!_(p))return;const B=l.get("path")||"/",j=F(B,p);a({path:j,fileId:null});return}if(!_(p)){if((window.location.pathname||"").includes("/apps/"+y+"/public/"))return;me(o)&&window.location.assign(H(o,""));return}const g=l.get("path")||"/",E=F(g,p);window.location.assign(H(o,E))},A=()=>{const o=(window.location.pathname||"")+"?"+(window.location.search||"");o!==n&&(n=o,L()&&u(),dt(),c(),m(),C(),r.scheduleRefresh(0))},x=()=>{const o="EtherpadNextcloudRouteWatchInstalled";if(window.OCA&&window.OCA[o]===!0)return;window.OCA&&(window.OCA[o]=!0);const l=()=>window.setTimeout(A,0);window.addEventListener("popstate",l),window.addEventListener("hashchange",l);const p=g=>{const E=window.history&&window.history[g];typeof E=="function"&&(window.history[g]=function(...B){const j=E.apply(this,B);return l(),j})};p("pushState"),p("replaceState")},S=()=>{i||(i=!0,x(),A(),v(),N(),r.installObserver(),r.scheduleRefresh(200))};document.readyState==="loading"?document.addEventListener("DOMContentLoaded",S,{once:!0}):S()})();
 //# sourceMappingURL=etherpad_nextcloud-files-main.mjs.map
